+++ YACReaderLibrary/comic_vine/model/volume_search_query.h
@@ -1,0 +1,12 @@
+#ifndef VOLUME_SEARCH_QUERY_H
+#define VOLUME_SEARCH_QUERY_H
+
+#include <QString>
+
+struct VolumeSearchQuery {
+    QString volume;
+    int page;
+    bool exactMatch;
+};
+
+#endif // VOLUME_SEARCH_QUERY_H