+++ YACReaderLibrary/comic_vine/search_single_comic.h
@@ -1,7 +1,9 @@
 #ifndef SEARCH_SINGLE_COMIC_H
 #define SEARCH_SINGLE_COMIC_H
 
-#include <QWidget>
+#include <QtWidgets>
+
+#include "scraper_checkbox.h"
 
 class ScraperLineEdit;
 
@@ -10,14 +12,17 @@ class SearchSingleComic : public QWidget
     Q_OBJECT
 public:
     SearchSingleComic(QWidget *parent = nullptr);
-    QString getVolumeInfo();
+    QString getVolumeInfo() const;
+    bool getExactMatch() const { return exactMatchCheckBox->isChecked(); }
+    void setVolumeInfo(const QString &volume);
     QString getComicInfo();
     int getComicNumber();
     void clean();
 
 private:
-    ScraperLineEdit *titleEdit;
-    ScraperLineEdit *numberEdit;
+    // ScraperLineEdit *titleEdit;
+    // ScraperLineEdit *numberEdit;
     ScraperLineEdit *volumeEdit;
+    ScraperCheckBox *exactMatchCheckBox;
 };
 #endif // SEARCH_SINGLE_COMIC_H