+++ YACReaderLibrary/comic_vine/search_volume.cpp
@@ -1,6 +1,7 @@
 #include "search_volume.h"
 
 #include "scraper_lineedit.h"
+#include "scraper_checkbox.h"
 
 #include <QLabel>
 #include <QVBoxLayout>
@@ -12,12 +13,17 @@ SearchVolume::SearchVolume(QWidget *parent)
     label->setStyleSheet("QLabel {color:white; font-size:12px;font-family:Arial;}");
 
     volumeEdit = new ScraperLineEdit(tr("Series:"));
+    volumeEdit->setClearButtonEnabled(true);
+
+    exactMatchCheckBox = new ScraperCheckBox(tr("Use exact match search. Disable if you want to find volumes that match some of the words in the name."), this);
+    exactMatchCheckBox->setChecked(true);
 
     QVBoxLayout *l = new QVBoxLayout;
 
     l->addSpacing(35);
     l->addWidget(label);
     l->addWidget(volumeEdit);
+    l->addWidget(exactMatchCheckBox);
     l->addStretch();
 
     l->setContentsMargins(0, 0, 0, 0);